refactor pod `AeCiphertext` into a separate submodule
