+++ zk-token-sdk/src/pod/pod.rs
@@ -99,34 +99,6 @@ pub struct RangeProof256(pub [u8; 800]);
 unsafe impl Zeroable for RangeProof256 {}
 unsafe impl Pod for RangeProof256 {}
 
-/// Serialization for AeCiphertext
-#[derive(Clone, Copy, PartialEq, Eq)]
-#[repr(transparent)]
-pub struct AeCiphertext(pub [u8; 36]);
-
-// `AeCiphertext` is a Pod and Zeroable.
-// Add the marker traits manually because `bytemuck` only adds them for some `u8` arrays
-unsafe impl Zeroable for AeCiphertext {}
-unsafe impl Pod for AeCiphertext {}
-
-impl fmt::Debug for AeCiphertext {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{:?}", self.0)
-    }
-}
-
-impl fmt::Display for AeCiphertext {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", BASE64_STANDARD.encode(self.0))
-    }
-}
-
-impl Default for AeCiphertext {
-    fn default() -> Self {
-        Self::zeroed()
-    }
-}
-
 // TODO: refactor this code into the instruction module
 #[derive(Clone, Copy, Pod, Zeroable)]
 #[repr(C)]
